requirements.txt
aiogram==3.1.1
aiohttp~=3.8.5
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi>=2023.7.22
charset-normalizer==2.1.0
dnspython==2.2.1
flake8==4.0.1
frozenlist==1.3.3
idna==3.3
mccabe==0.6.1
multidict==6.0.2
# openai==0.19.0
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
yarl==1.8.2
